feat: Created new model, updates to run on local mac

+++ torchtitan/models/common/moe.py
@@ -267,11 +267,11 @@ class TokenChoiceTopKRouter(Module):
 
         # group tokens together by expert indices from 0 to num_experts and pass that to experts forward
         num_tokens_per_expert = torch.histc(
-            selected_experts_indices.view(-1),
+            selected_experts_indices.view(-1).float(),
             bins=self.num_experts,
             min=0,
             max=self.num_experts,
-        )
+        ).to(torch.int32)
 
         return top_scores, selected_experts_indices, num_tokens_per_expert
 
@@ -310,11 +310,11 @@ class TokenReorderer(Module):
         """
         # group tokens together by expert indices from 0 to num_experts and pass that to experts forward
         num_tokens_per_expert = torch.histc(
-            selected_experts_indices.view(-1),
+            selected_experts_indices.view(-1).float(),
             bins=self.num_experts,
             min=0,
             max=self.num_experts,
-        )
+        ).to(torch.int32)
 
         # Reorder the token indices to match the order of the experts
         # token_indices_experts_sorted shape (bs*slen*top_k,)
@@ -471,7 +471,7 @@ class MoE(Module):
         out = deterministic_scatter_add(
             out,
             token_indices_experts_sorted.reshape(-1, 1).expand(-1, dim),
-            routed_output,
+            routed_output.to(out.dtype),
         )
         out = out.reshape(bs, slen, dim)
         return out
